Refactor statistics widget and adding generic method
